Move delete cluster button from details main page to the resize tab
